getting current batch id to work

+++ BeerProductionAPI/BeerProductionAPI/LiveRelevantData.cs
@@ -44,7 +44,7 @@ namespace BeerProductionAPI
         [DataMember]
         public int CurrentState { get; set; }
         [DataMember]
-        public float CurrentBatchID { get; set; }
+        public int CurrentBatchID { get; set; }
         /*
         [DataMember]
         public Dictionary<int, TimeSpan> StateDictionary { get; set; }
@@ -52,7 +52,7 @@ namespace BeerProductionAPI
         public LiveRelevantData(float temperature, float humidity, float vibration,
             float actualMachineSpeed, int producedProducts, 
             int defectProducts, float barley, float hops, float malt, float wheat, 
-            float yeast, ushort maintainenceMeter, int currentState, float batchID, float batchSize, float currentBatchID)
+            float yeast, ushort maintainenceMeter, int currentState, float batchID, float batchSize, int currentBatchID)
         {
             Temperature = temperature;
             Humidity = humidity;
